% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{add_component_column}
\alias{add_component_column}
\title{add_component_column}
\usage{
add_component_column(table, add = c(), exclude = c())
}
\arguments{
\item{table}{a table to rename}

\item{add}{a prefix containing the component e.g. 'AAPL' for Apple}

\item{exclude}{a vector of column names to exclude}
}
\value{
table with modified columns
}
\description{
Rename a column by adding a prefix
Antonym of "remove_component_column"
}
