% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{calc_deFacto_weights}
\alias{calc_deFacto_weights}
\title{calc_deFacto_weights}
\usage{
calc_deFacto_weights(
  time_series,
  price.type = c("Close", "Open", "High", "Low", "Volume", "Adjusted"),
  rename = FALSE
)
}
\arguments{
\item{time_series}{A table containing the time series of prices for each component
and their index parameters}

\item{price.type}{A string indicating which price to use}

\item{rename}{A boolean indicating if column "Weights" should be renamed to "Weights.{Price.Type}"
e.g., "Weights.Close"}
}
\value{
A new table containing de-facto weights  per unit of time (e.g. 'day')
}
\description{
Calculate deFactoWeights from individual component values and index parameters
}
