% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{calc_index}
\alias{calc_index}
\title{calc_index}
\usage{
calc_index(
  time_series,
  share.col = Shares,
  price.col = Price,
  price.type = c("Close", "Open", "High", "Low", "Volume", "Adjusted"),
  rename = c(FALSE, TRUE)
)
}
\arguments{
\item{time_series}{A table containing the time series of prices for each component
and their index parameters}

\item{share.col}{Name of column (unquoted) containing the prices (default: Price)}

\item{price.type}{A string indicating which price to use}

\item{rename}{A boolean indicating if column "Index" should be renamed to "Index.{Price.Type}"
e.g., "Index.Close"}
}
\value{
A new table containing index values per unit of time (e.g. 'day')
}
\description{
Calculate index price from individual component values
}
