% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{calc_several_index}
\alias{calc_several_index}
\title{calc_several_index}
\usage{
calc_several_index(
  timeseries,
  price.type = c("Close", "Open", "High", "Low", "Volume", "Adjusted")
)
}
\arguments{
\item{timeseries}{Time series of prices in long format}

\item{price.type}{For which price to compute the index}
}
\value{
A timeseries of index prices
}
\description{
Compute index values for one or more price types (e.g., Opening price, High prices...)
Expand function calc_index() to several price types
}
