% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{checkpoint}
\alias{checkpoint}
\title{checkpoint}
\usage{
checkpoint(data, load = FALSE)
}
\arguments{
\item{data}{A variable to save}

\item{load}{A boolean indicating whether to save, or load the data}
}
\value{
Assign to Global environment
}
\description{
Create a checkpoint with the data and
can load the data from the previous checkpoint

Use: Saving a copy before testing code
}
