% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{compute_across}
\alias{compute_across}
\title{compute_across}
\usage{
compute_across(..., .fun)
}
\arguments{
\item{...}{objects to be concatenated}

\item{.fun}{a function to apply}
}
\value{
A single value
}
\description{
Apply a function across all values provided
}
