% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{extract_price}
\alias{extract_price}
\title{extract_price}
\usage{
extract_price(date, price, .fun, message = FALSE, return.date = FALSE)
}
\arguments{
\item{date}{A period of time, matching price}

\item{price}{A time series of price, matching date}

\item{.fun}{A function to apply for extracting a date
'min' returns the first day, 'max' returns the last day}

\item{message}{A boolean specifying whether to display a summary or not}

\item{return.date}{A boolean specifying whether to extract the date instead}
}
\value{
A single value: a price corresponding to a specific date (or a date if 'date' is TRUE)
}
\description{
Extract price for first, last, and middle day of the period
}
