% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{merge_data_parameters}
\alias{merge_data_parameters}
\title{merge_data_parameters}
\usage{
merge_data_parameters(
  time_series,
  to_merge_with,
  by_col = NULL,
  is_long = TRUE
)
}
\arguments{
\item{time_series}{A table containing the time series of prices for each component}

\item{to_merge_with}{Table containing index parameters (e.g. number of shares)}

\item{by_col}{Which column to merge with}

\item{is_long}{A boolean indicating the table is in long or wide format.
If wide, will convert the table to long format}
}
\value{
A new table containing index values per unit of time (e.g. 'day')
}
\description{
Calculate index price from individual component values
}
