% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{periodify}
\alias{periodify}
\title{periodify}
\usage{
periodify(
  timeseries,
  date_col,
  period = c("day", "week", "month", "year"),
  create_group = FALSE
)
}
\arguments{
\item{timeseries}{a timeseries table containing a variable for time}

\item{date_col}{name of the column containing dates}

\item{period}{a vector of characters indicating which time components to extract (pick one or more of the
following options: day, week, month, year}

\item{create_group}{If TRUE, combines extracted periods for future grouping by period}
}
\value{
the timeseries table with new columns for the extracted time components
}
\description{
Extract time components from a time series and add it to the table containing the time series.
}
\details{
This helper function can be used in combination with a grouping and an aggregation function to calculate weekly, monthly, or yearly values
}
