% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{pivot_prices_longer}
\alias{pivot_prices_longer}
\title{pivot_prices_longer}
\usage{
pivot_prices_longer(data, symbols, separate = FALSE)
}
\arguments{
\item{data}{A wide table}

\item{symbols}{Symbols of financial components}

\item{separate}{A boolean, if TRUE, separate the component from the price type (e.g. 'AAPL' and 'Close' price)}
}
\value{
A tibble in wider format
}
\description{
Take a wide table of financial data and pivot the table
}
\concept{pivot_prices_wider}
