% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{pivot_prices_wider}
\alias{pivot_prices_wider}
\title{pivot_prices_wider}
\usage{
pivot_prices_wider(data, nameS, valueS)
}
\arguments{
\item{data}{A long table}

\item{nameS}{A vector of names for the column or columns containing column identifiers}

\item{valueS}{A vector of names containing the variable(s)}
}
\value{
A tibble in wider format
}
\description{
Take a long table of financial data and pivot the table by reuniting Component and Price types if needed
}
\concept{pivot_prices_longer}
