% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{remove_component_column}
\alias{remove_component_column}
\title{remove_component_column}
\usage{
remove_component_column(table, remove = c())
}
\arguments{
\item{table}{the table to rename}

\item{remove}{a vector of length containing the pattern}
}
\value{
the table with new column names
}
\description{
Rename a column by removing the component pattern
}
