% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{select_partial_colnames}
\alias{select_partial_colnames}
\title{select_partial_colnames}
\usage{
select_partial_colnames(data, column_to_keep)
}
\arguments{
\item{data}{a table}

\item{column_to_keep}{a vector of partial column names}
}
\value{
a vector containing the index of columns
to select based on partial names
}
\description{
Select columns of a table using partial names,
e.g. select "AAPL.Close" from "Close"
}
