% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{tib_to_tsib}
\alias{tib_to_tsib}
\title{tib_to_tsib}
\usage{
tib_to_tsib(
  tibble_ts,
  key_ts = c(),
  index_ts = c(),
  keep_price = c("All", "Open", "High", "Low", "Volume", "Adjusted", "Close"),
  wide = TRUE
)
}
\arguments{
\item{tibble_ts}{a tibble containing a time series}

\item{key_ts}{a character vector containing the identifying keys for the different time series}

\item{index_ts}{a character vector containing the time indexes}

\item{keep_price}{a vector character of price names to select}

\item{wide}{Is the table in wide format? Default: TRUE. If FALSE, will pivot the table}
}
\value{
a tsibble, a.k.a. "time series tibble"
}
\description{
Coerce tibble to tsibble for time series
}
