% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_wrangling.R
\name{xts_finance_to_tibble}
\alias{xts_finance_to_tibble}
\title{xts_finance_to_tibble}
\usage{
xts_finance_to_tibble(symbol, from_date, to_date)
}
\arguments{
\item{symbol}{A character vector containing the Symbols of a component,
e.g. 'AAPL' for Apple}

\item{from_date}{Start of the time period}

\item{to_date}{End of the time period}
}
\value{
A tibble containing the time series associated to these symbols
}
\description{
Download financial xts data and coerce to tibble
}
